the change speed functions should now work as intended

+++ maze_car/motor_shield.h
@@ -61,8 +61,8 @@ public:
   void set_speed(uint8_t speed_m1, uint8_t speed_m2, uint8_t speed_m3, uint8_t speed_m4);
   void set_speed(uint8_t* speed);
   void set_speed(int8_t motor, uint8_t speed_m);
-  const uint8_t* change_speed(int8_t speed_m1, int8_t speed_m2, int8_t speed_m3, int8_t speed_m4);
-  void change_speed(int8_t motor, int8_t speed_m);
+  const uint8_t* change_speed(int16_t speed_m1, int16_t speed_m2, int16_t speed_m3, int16_t speed_m4);
+  void change_speed(int8_t motor, int16_t speed_m);
   void update_speed(void);
 
   void change_motor_direction(uint8_t dir1, uint8_t dir2, uint8_t dir3, uint8_t dir4);
